feat: handle WHIP endpoints that does not support trickle ICE
Added support in the client SDK for the application to force not
to use trickle ICE. Useful when you know a WHIP endpoint does not support
it and does not signal it upfront in any way.
If the force option is not set it will try to figure the support out
by checking the allowed-methods header and if PATCH is not included
it will assume that trickle ICE is not supported.